refactor: Movendo lógica para custom hooks
