+++ frontend/src/pages/Home/index.tsx
@@ -1,37 +1,14 @@
 import { Player } from "@lottiefiles/react-lottie-player";
-import { useNavigate } from "react-router-dom";
-import bookImage from "../../assets/bookimage.jpeg";
+import { useHome } from './useHome';
+
 import styles from "./styles.module.scss";
 
 export const Home = () => {
-  const navigate = useNavigate();
-
-  const spamPhrases = [
-    "O sucesso em forma de livro!",
-    "Os melhores livros para enriquecer sua mente!",
-    "Os mais vendidos e amados, agora ao seu alcance!",
-    "A porta de entrada para um mundo de conhecimento!",
-    "Best sellers para alimentar sua mente!",
-    "Descubra histórias incríveis nos melhores livros!",
-    "Descubra por que esses livros são os mais vendidos!",
-    "A leitura que transforma a sua vida!",
-    "Os best sellers que não podem faltar na sua estante!",
-    "Leia os melhores livros e amplie seu universo!",
-    "Os livros mais populares estão aqui!",
-    "Sua dose diária de inspiração está nos melhores livros!",
-    "Best sellers para todas as idades e gostos!",
-    "Aqui, você encontra os melhores livros para transformar seu dia!",
-    "Os livros que estão conquistando o mundo!",
-    "Os melhores livros, a chave para um mundo mais rico!",
-    "Leia os best sellers e faça parte da história!",
-    "Leia os melhores livros e viva novas aventuras a cada página!",
-    "Aqui você encontra os best sellers que vão te surpreender!",
-    "Os melhores livros, uma viagem sem sair do lugar!",
-  ];
+  const { navigate, spamPhrases, books } = useHome();
 
   return (
     <div>
-      <div className={styles["container-1"]}>
+      <section className={styles["container-1"]}>
         <div className={styles.subcontainer}>
           <p className={styles.title}>Descubra novas histórias!</p>
           <span>
@@ -48,7 +25,7 @@ export const Home = () => {
           loop
           autoplay
         />
-      </div>
+      </section>
 
       <div className={styles["infinite-animation"]}>
         <p>
@@ -65,36 +42,19 @@ export const Home = () => {
         </p>
       </div>
 
-      <div className={styles["container-2"]}>
+      <section className={styles["container-2"]}>
         <h1 className={styles.title}>Livros mais vendidos</h1>
         <div className={styles.subcontainer}>
-          <img
-            src={bookImage}
-            alt="capa do livro"
-            onClick={() => navigate("/books/1")}
-          />
-          <img
-            src={bookImage}
-            alt="capa do livro"
-            onClick={() => navigate("/books/1")}
-          />
-          <img
-            src={bookImage}
-            alt="capa do livro"
-            onClick={() => navigate("/books/1")}
-          />
-          <img
-            src={bookImage}
-            alt="capa do livro"
-            onClick={() => navigate("/books/1")}
-          />
-          <img
-            src={bookImage}
-            alt="capa do livro"
-            onClick={() => navigate("/books/1")}
-          />
+          {books.map((book, i) => (
+            <img
+            key={i}
+            src={book.image}
+            alt={book.alt}
+            onClick={() => navigate(`/books/${book.id}}`)}
+            />
+          ))}
         </div>
-      </div>
+      </section>
     </div>
   );
 };
